requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
falcon==2.0.0             # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via requests
influxdb==5.3.0           # via -r requirements.in
msgpack==0.6.1            # via influxdb
paho-mqtt==1.5.1          # via -r requirements.in
python-dateutil==2.8.1    # via influxdb
python-decouple==3.3      # via -r requirements.in
pytz==2020.1              # via influxdb
requests==2.24.0          # via influxdb
six==1.15.0               # via influxdb, python-dateutil
urllib3==1.25.10          # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
